安全升级、文章相关功能调整、jwt生成逻辑修改

+++ src/main/java/com/linghang/we_talk/controller/ArticleController.java
@@ -1,38 +1,3 @@
-//package com.linghang.we_talk.controller;
-//
-//import com.linghang.we_talk.entity.Comment;
-//import com.linghang.we_talk.utils.Result;
-//import io.swagger.v3.oas.annotations.Operation;
-//import io.swagger.v3.oas.annotations.tags.Tag;
-//import org.springframework.web.bind.annotation.PostMapping;
-//import org.springframework.web.bind.annotation.RequestBody;
-//import org.springframework.web.bind.annotation.RequestMapping;
-//import org.springframework.web.bind.annotation.RestController;
-//
-//@Tag(name = "articleController",description = "关于文章的一些CRUD服务")
-//@RestController
-//@RequestMapping("/article")
-//public class ArticleController {
-//    //预想：
-//    //1.添加文章
-//    //2.删除文章
-//    //3.我的文章
-//    //4.我赞过的
-//    //5.获取最近指定数目文章 --- 分页查询？整一个吧
-//    //5.5关键词搜索
-//    //6.发表评论
-//    //-------归于adminController，划分职责。
-//    //7.被举报的dirty文章
-//    //8.裁判删除或者归于正常dirty文章
-//
-//    @Operation(summary = "添加评论")
-//    @PostMapping("/addComment")
-//    public Result<?> addComment(@RequestBody Comment comment){
-//
-//        return Result.succeed("true");
-//    }
-//}
-
 package com.linghang.we_talk.controller;
 
 import com.linghang.we_talk.DTO.ArticleCreateRequest;
@@ -48,12 +13,12 @@ import jakarta.annotation.Resource;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Objects;
 
 @Slf4j
 @Validated
@@ -73,10 +38,12 @@ public class ArticleController {
     /**
      * 创建文章
      */
-    @Operation(summary = "创建文章")
+    @Operation(summary = "创建文章",description = "分类ID是需要你手动传的，为空的话数据库里没有默认值。")
     @PostMapping("/create")
     public Result<?> createArticle(
-            @Valid @RequestBody ArticleCreateRequest request) {
+            @Valid @RequestBody ArticleCreateRequest request,@RequestAttribute("userid") String userid) {
+        //安全升级
+        request.setUserId(Long.parseLong(userid));
         Long articleId = articleService.createArticle(request);
         return Result.succeed(articleId);
     }
@@ -85,11 +52,15 @@ public class ArticleController {
      * 更新文章
      */
     @Operation(summary = "更新文章")
-    @GetMapping("/upt/{id}")
+    @PostMapping("/upt")
     public Result<?> updateArticle(
-            @PathVariable Long id,
-            @Valid @RequestBody ArticleUpdateRequest request) {
-        request.setId(id);
+            @Valid @RequestBody ArticleUpdateRequest request,@RequestAttribute("userid") String userid) {
+
+        ArticleVO articleDetail = articleService.getArticleDetail(request.getId());
+        Long userId = Long.parseLong(userid);
+
+        if(!Objects.equals(articleDetail.getUserId(), userId)) return Result.error("非法更新");
+
         articleService.updateArticle(request);
         return Result.succeed();
     }
@@ -101,9 +72,20 @@ public class ArticleController {
     @GetMapping("/del/{id}")
     public Result<?> deleteArticle(
             @PathVariable Long id,
-            @RequestHeader("X-User-Id") Long userId) {
-        articleService.deleteArticle(id, userId);
-        return Result.succeed();
+            @RequestAttribute("userid") String userid) {
+
+        try {
+            ArticleVO articleDetail = articleService.getArticleDetail(id);
+            Long userId = Long.parseLong(userid);
+
+            if(!Objects.equals(articleDetail.getUserId(), userId)) return Result.error("非法删除");
+
+            articleService.deleteArticle(id, userId);
+
+            return Result.succeed();
+        } catch (NumberFormatException e) {
+            return  Result.error("删除失败，请检查参数");
+        }
     }
 
     /**
@@ -112,14 +94,14 @@ public class ArticleController {
     @Operation(summary = "根据ID获取文章详情")
     @GetMapping("/{id}")
     public Result<ArticleVO> getArticleDetail(@PathVariable Long id) {
-        ArticleVO article = null;
+
         try {
-            article = articleService.getArticleDetail(id);
+            ArticleVO article = articleService.getArticleDetail(id);
+            return Result.succeed(article);
         } catch (Exception e) {
             log.info("文章不存在或者被软删除，ID：{},{}",id,e.getMessage());
             return Result.error("文章不存在");
         }
-        return Result.succeed(article);
     }
 
     /**
@@ -140,7 +122,7 @@ public class ArticleController {
      * 获取用户文章列表
      */
     @Operation(summary = "获取用户所有文章")
-    @GetMapping("/user/{userId}")
+    @GetMapping("/user/{userId}")//无需安全升级，为公共接口
     public Result<List<ArticleVO>> getUserArticles(@PathVariable Long userId) {
         List<ArticleVO> articles = articleService.getUserArticles(userId);
         return Result.succeed(articles);
@@ -157,7 +139,7 @@ public class ArticleController {
             String username =  jwtUtil.validateToken(accessToken);
 
             User user = userService.getUserByName(username);
-            Integer userId = user.getId();
+            Long userId = user.getId();
 
             if (isLike) {
                 articleService.likeArticle(id, userId);
@@ -179,7 +161,7 @@ public class ArticleController {
         String username =  jwtUtil.validateToken(accessToken);
 
         User user = userService.getUserByName(username);
-        Integer userId = user.getId();
+        Long userId = user.getId();
 
         return Result.succeed(articleService.checkLiked(articleId,userId));
     }
